==> SNA and GNNs/Code/3. Transactions Graph.ipynb <==
# 3. Transactions Graph

import pandas as pd
import numpy as np
import dgl
import torch as th
from sklearn.preprocessing import StandardScaler

# Import Data

# import new data after feature engineering
wires = pd.read_csv('wires.csv').iloc[:, 1:]
emts = pd.read_csv('emts.csv').iloc[:, 1:]
cash = pd.read_csv('cash.csv').iloc[:, 1:]
cust_info = pd.read_csv('cust_info.csv').iloc[:, 1:]
detail_cust = pd.read_csv('detailed_cust_info.csv').iloc[:, 1:]
ext_info = pd.read_csv('external_info.csv').iloc[:, 1:]

# convert the amount in cash df to type float
cash['amount'] = cash['amount'].astype(float)

# Defining Transaction Relationships

We will be making a directed heterogeneous graph to model the relationships between various wire, emt and cash transactions.
- Find transactions that **share a user between the transactions**
    - e.g. each transaction will have a sender and receiver
    - potential shared users include (we will not focus on those who have the same sender or the same receiver):
        - **receiver transaction A = sender transaction B**
- find those shared transactions who **have a similar transaction amount**
    - Set a threshold for what counts as similar. This is vary depending on the transactions
        - If user sends/receives money, then we **assume that amount received => amount sent**
        - If amount received => amount sent --> we will **create an edge between these two transactions**

Note: There will be no edge features for this graph, only node features.

### Graph Nodes and Edges

For the heterogeneous directed transactions graph:

Nodes:
- wire transactions
- emt transactions
- cash transactions
    
Edges
- Sender/receiver: transactions need to share at least 1 person that is the same 
    - receiver transaction A = sender transaction B
- Transaction amount: amount received => amount sent

### **Canonical Edge Types**:</br>

Node: **edge abbreviations** are in the format of **source-shares-destination**
- For example: wire-share-wire = 'wsw'

**Between `wire`/`emt` transactions**</br>
- ('wire', 'wsw', 'wire')</br>
- ('wire', 'wse', 'emt')</br>
- ('emt', 'ese', 'emt')</br>


**Between `wire`/`emt` --> `Cash Withdrawal` transactions**</br>
- ('wire', 'wsc', 'cash') --> receive money through transfer then withdraw the money as cash to send elsewhere</br>
- ('emt', 'esc', 'cash')</br>

**Between `Cash Deposits` --> `wire`/`emt` transactions**
- ('cash', 'csw', 'wire') --> deposit cash from someone then trasfer the money elsewhere</br>
- ('cash', 'cse', 'emt')</br>

**Between `Cash withdrawals` --> `Cash deposits`**
- ('cash', 'csc', 'cash') --> match on amounts with withdrawal --> deposit relationship for different people</br>

Send and receive relations summary:
- receive wire/emt --> send wire/emt
- receive wire/emt --> deposit cash
- withdrawal cash --> send wire/emt

# Get node features for wires

wires.head()

wire1 = wires[['trxn_id', 'sender_global_id', 'rec_global_id', 'wire value']].rename(columns={'wire value':'amount'})
wire1

# Get node features for emts

emts.head()

emt1 = emts[['trxn_id', 'sender_global_id', 'rec_global_id', 'emt value']].rename(columns={'emt value':'amount'})
emt1

# Get node features for cash 

cash1 = cash[['trxn_id', 'Global_id', 'amount', 'type_deposit']]
cash1

# Get COO Tensors for Edges

### Create a map for all transactions

# Get list of transaction ids from the three transaction tables
ids = []
w_ids = list(wires['trxn_id'].values)
e_ids = list(emts['trxn_id'].values)
c_ids = list(cash['trxn_id'].values)

# append the ids to the ids list
ids += w_ids + e_ids + c_ids

# get global_ids for indexes of customers
indexes = [ x for x in range(len(ids)) ]

# create a dictionary for the mapped indexes
t_map = {gid:index for gid, index in zip(ids, indexes)}

checking = pd.DataFrame({'trxn_id':ids, 'index':indexes})
checking

# export the transaction map
checking.to_csv('trxn_map.csv')

#### Create Function to get edges and map indexes

We will use a exponential decay function to calculate the weights for the relationship.
When the difference between the amounts increase, the strong the relationship will be and the weight will move towards 1. We use an exponential decay function to model this relationship, since as the difference grows, the final value moves towards 1.

def calc_weightother(amt1, amt2):
    return 1 - np.exp(-abs(amt1 - amt2) / 100)

# create a function to track the number of times a customer is listed in the COO format
def count_participation(m, cash):
    """
    """
    
    # Get the columns that have global_id
    g_cols = list(m.filter(like='global_id').columns)
    
    if cash:
        # concat the columns that have ids into a single column
        ids = pd.concat([m[g_cols[0]], m['common']]).reset_index(drop=True)
    else:
        # concat the columns that have ids into a single column
        ids = pd.concat([m[g_cols[0]], m[g_cols[1]], m['common']]).reset_index(drop=True)

    # Store the customer participation 
    id_counts = ids.value_counts().reset_index()
    
    # Rename the columns
    id_counts = id_counts.rename(columns={'index': 'Global_id'})
    
    return id_counts

# Create a function to get the edges between two transactions
def get_edges(receiver_df, sender_df, withdrawal, deposit):
    """
    
        Parameters:
            - receiver_df (DataFrame): dataframe containing transactions with incoming money
            - sender_df (DataFrame): dataframe containing transactions with outgoing money
            - withdrawal (Boolean): 1 if contains a withdrawal trxn
            - deposit (Boolean): 1 if contains a deposit trxn
    """
    
    # Set cash to false for pariticipation function
    cash = False
    
    # Rename the sender in sender_df common and rename the receiver in receiver_df common to merge
    if deposit:
        cash = True
        receiver_df = receiver_df.rename(columns={'Global_id': 'common'})
    else:
        receiver_df = receiver_df.rename(columns={'rec_global_id': 'common'})
    
    if withdrawal:
        cash = True
        sender_df = sender_df.rename(columns={'Global_id': 'common'})
    else:
        sender_df = sender_df.rename(columns={'sender_global_id': 'common'})
    
    # Merge the two transaction dataframes on the common sender column
    m = pd.merge(receiver_df, sender_df, on='common')
    
    # Store the participation counts for each customer in this network
    participation = count_participation(m, cash)
    
    # filter merged for where amount_x < amount_y
    filtered = m[m['amount_y'] <= m['amount_x']].rename(columns={'trxn_id_x': 'Source', 'trxn_id_y': 'Destination'})
    
    # Calculate the weight of the edge based on the amount columns
    filtered['Weight'] = calc_weightother(filtered['amount_x'], filtered['amount_y'])
    
    # Get the COO for the transactions
    coo = filtered[['Source', 'Destination', 'Weight']]
    
    return coo, participation

# create a function to map edges with the corresponding index values in the node features dataframe
def map_edges(df):
    """
    """
    edges_df = pd.DataFrame()
    edges_df['Source'] = df['Source'].map(t_map)
    edges_df['Destination'] = df['Destination'].map(t_map)
    edges_df['Weight'] = df['Weight']
    
    return edges_df

# create a function to store the edges dataframe as a tuple with tensors
def get_tensors(edges):
    """
    """
    
    # transpose the dataframe to get the format for the tensors
    values = edges.values.transpose()
    
    # convert the values to tensors
    u = th.tensor(values[0].copy())
    v = th.tensor(values[1].copy())
    
    # return a tuple of node_tensors
    return (u, v)

# create copies of the transaction dataframes to find the edges
w1 = wire1.copy()
w2 = wire1.copy()
e1 = emt1.copy()
e2 = emt1.copy()

# Get cash deposits vs withdrawals
deposits = cash1[cash1['type_deposit'] == 1][['trxn_id', 'Global_id', 'amount']]
withdrawals = cash1[cash1['type_deposit'] == 0][['trxn_id', 'Global_id', 'amount']]

# Find send/receive transaction relations

### Extract COO for ('wire', 'wsw', 'wire')

These edges model the relationship where an individual received money through wire and then sent money through wire at a smaller amount than the amount they received. Source indicates the transaction id where they received the wire transaction and destination is the transaction where they sent a smaller amount.

# Get the edges for the edge type
wsw, wsw_part = get_edges(w1, w2, False, False)

# Apply the proper mapping to get the COO 
wsw_edges = map_edges(wsw)
wsw_edges

# Show graph participation
wsw_part

# get the node-tensors for these edges
wsw_ten = get_tensors(wsw_edges)
wsw_ten

### Extract COO for ('wire', wse', 'emt')

These edges model the relationship where an individual received money through wire and then sent money through emt at a smaller amount than the amount they received. Source indicates the transaction id where they received the wire transaction and destination is the transaction where they sent a smaller amount using an emt.

# Get the coo format
wse, wse_part = get_edges(w1, e2, False, False)

# Apply the proper mapping to get the COO 
wse_edges = map_edges(wse)
wse_edges

# get the node-tensors for these edges
wse_ten = get_tensors(wse_edges)
wse_ten

### Extract COO for ('emt', 'same amt', 'wire')

# Get the coo format
esw, esw_part = get_edges(e1, w2, False, False)

# Apply the proper mapping to get the COO 
esw_edges = map_edges(esw)
esw_edges

# get the node-tensors for these edges
esw_ten = get_tensors(esw_edges)
esw_ten

### Extract COO for ('emt', 'same amt', 'emt')

Due to the large amount of emts, we will also filter the transactions after standardizing the transaction amount values. We do this by standardizing the transaction amounts then filtering for those that amore than 3 standard deviations away from the mean.

# Set the receiver df and sender df 
receiver_df = e1.rename(columns={'rec_global_id': 'common'})
sender_df = e2.rename(columns={'sender_global_id': 'common'})

# merge the dataframe on the common customer
m = pd.merge(receiver_df, sender_df, on='common')

# count participation of customers in this network
ese_part = count_participation(m, False)

# standardize the amounts 
emt_scalerx = StandardScaler()
emt_scalery = StandardScaler()

# Select amount column and reshape
amtx = m['amount_x'].values.reshape(-1, 1)
amty = m['amount_y'].values.reshape(-1, 1)

# Fit Transform amounts
m['amt_x_std'] = emt_scalerx.fit_transform(amtx)
m['amt_y_std'] = emt_scalery.fit_transform(amty)

# filter for outliers
filtered = m[(m['amt_x_std'] > 3) | (m['amt_x_std'] < -3) | (m['amt_y_std'] > 3) | (m['amt_y_std'] < -3)]

# Calculate the weight of the edge based on the amount columns
filtered['Weight'] = calc_weightother(filtered['amount_x'], filtered['amount_y'])

# rename
filtered = filtered.rename(columns={'trxn_id_x': 'Source', 'trxn_id_y': 'Destination'})

# Get the COO for the transactions
ese = filtered[['Source', 'Destination', 'Weight']]

# Apply the proper mapping to get the COO 
ese_edges = map_edges(ese)
ese_edges

# get the node-tensors for these edges
ese_ten = get_tensors(ese_edges)
ese_ten

# Find receive wire/emt --> withdrawal cash transaction relations

These COOs will represent relations where an individual receives a wire or emt transaction and also withdraws cash for a similar amount. Can represent the flow of money where some one receives money through wire/emt, then withdraws cash to send elsewhere.

### Extract COO for ('wire', 'same amt', 'cash')

# Get the coo format
wsc, wsc_part = get_edges(w1, withdrawals, True, False)

# Apply the proper mapping to get the COO 
wsc_edges = map_edges(wsc)
wsc_edges

# get the node-tensors for these edges
wsc_ten = get_tensors(wsc_edges)
wsc_ten

### Extract COO for ('emt', 'same amt', 'cash')

# Get the coo format
esc, esc_part = get_edges(e1, withdrawals, True, False)

# Apply the proper mapping to get the COO 
esc_edges = map_edges(esc)
esc_edges

# get the node-tensors for these edges
esc_ten = get_tensors(esc_edges)
esc_ten

# Find deposit --> send wire/emt transaction relations

These COOs will represent relations where an individual deposits cash they received from somewhere, then sends a wire/emt transfer of a smaller amount to someone else. It can represent the flow of money where some one receives money through cash and deposits it, then sends money through wire/emt to somewhere else.

### Extract COO for ('cash', 'same amt', 'wire')

# Get the coo format
csw, csw_part = get_edges(deposits, w2, False, True)

# Apply the proper mapping to get the COO 
csw_edges = map_edges(csw)
csw_edges

# get the node-tensors for these edges
csw_ten = get_tensors(csw_edges)
csw_ten

### Extract COO for ('cash', 'same amt', 'emt')

# Get the coo format
cse, cse_part = get_edges(deposits, e2, False, True)

# Apply the proper mapping to get the COO 
cse_edges = map_edges(cse)
cse_edges

# get the node-tensors for these edges
cse_ten = get_tensors(cse_edges)
cse_ten

# Find cash --> cash transaction relations

This COO represents the flow of cash through from individuals that withdraw cash to individuals who deposit cash for the same amount. These edges can then help us detect cash transactions between individuals.

Since there are many cash transactions that share the same amount, we will filter for only those transactions that are unusually large or small. This is measured by standardizing the transaction amount then finding those that are greater than 3 standard deviations from the mean.

Since we are merging on the amount, the weight will be 1 for all of these edges.

# Create a function to get the edges for cash transactions
def cash_edges(withdrawals, deposits):
    """    
        Parameters:
            - receiver_df (DataFrame): dataframe containing transactions with incoming money
            - sender_df (DataFrame): dataframe containing transactions with outgoing money
    """
    # Merge the two transaction dataframes on the amount column
    m = pd.merge(withdrawals, deposits, on='amount').rename(columns={'trxn_id_x': 'Source', 'trxn_id_y': 'Destination'})
    
    # Get customer participation
    g_cols = list(m.filter(like='Global_id').columns)
    ids = pd.concat([m[g_cols[0]], m[g_cols[1]]]).reset_index(drop=True)
    id_counts = ids.value_counts().reset_index()
    id_counts = id_counts.rename(columns={'index': 'Global_id'})
    
    # Select amount column and reshape
    amts = m['amount'].values.reshape(-1, 1)
    
    # Create a new variable for the standardizing the amount
    cash_scaler = StandardScaler()
    m['std'] = cash_scaler.fit_transform(amts)
    
    # filter for amounts that are greater than 3 std from mean
    filtered = m[(m['std'] > 3) | (m['std'] < -3)]
    
    filtered.loc[:, 'Weight'] = 1
    
    # Get the COO for the transactions
    coo = filtered[['Source', 'Destination', 'Weight']]
    
    return coo, id_counts

### Extract COO for ('cash', 'same amt', 'cash')

# Get the coo format
csc, csc_part = cash_edges(withdrawals, deposits)

# Apply the proper mapping to get the COO 
csc_edges = map_edges(csc)
csc_edges

# get the node-tensors for these edges
csc_ten = get_tensors(csc_edges)
csc_ten

# Export customer participation dataframes

wsw_part.to_csv('wsw_part.csv')
wse_part.to_csv('wse_part.csv')
wsc_part.to_csv('wsc_part.csv')
ese_part.to_csv('ese_part.csv')
esc_part.to_csv('esc_part.csv')
esw_part.to_csv('esw_part.csv')
csc_part.to_csv('csc_part.csv')
csw_part.to_csv('csw_part.csv')
cse_part.to_csv('cse_part.csv')

# Export Subgraph Edges

wsw_edges.to_csv('wsw_edges.csv')
wse_edges.to_csv('wse_edges.csv')
wsc_edges.to_csv('wsc_edges.csv')
ese_edges.to_csv('ese_edges.csv')
esc_edges.to_csv('esc_edges.csv')
esw_edges.to_csv('esw_edges.csv')
csc_edges.to_csv('csc_edges.csv')
csw_edges.to_csv('csw_edges.csv')
cse_edges.to_csv('cse_edges.csv')

# Construct Graphs

Since we want to apply Louvain Modularity to find communities of transactions, we will create a graph for each unique edge type defined before. After creating these graphs, we will apply the louvain method to each to find communities within each graph of transactions. 

#!pip install leidenalg python-igraph

from igraph import Graph

### Concat all edges to form a single graph

# Create a list of edges to get 
edge_dfs = [wsw_edges, wse_edges, wsc_edges,
            ese_edges, esc_edges, esw_edges, 
            csc_edges, csw_edges, cse_edges]

# use a for loop to get all the tuples of edges an a list of corresponding weights
edges = []
weights = []

for df in edge_dfs:
    # Get the node pairs in a tuple
    eds = list(zip(df['Source'], df['Destination']))
    
    # Get the list of corresponding weights
    w = list(df['Weight'])
    
    # Add the edges to the edges list and ws to weights
    edges += eds
    weights += w

# Get the unique nodes in the graph
nodes = list(checking['trxn_id'].values)
len(nodes)

# Create a directed graph
g = Graph(directed=True)
g.add_vertices(nodes)
g.add_edges(edges)
g.es['weight'] = weights

### Check connections

# Print edge source, target, and weight to verify
for edge in g.es[:5]:
    print(f"Edge from {g.vs[edge.source]['name']} to {g.vs[edge.target]['name']} has weight {edge['weight']}")

wsw_edges.head()

checking[checking['trxn_id'] == 'SLBV29462341']

checking[checking['trxn_id'] == 'FAHV35141379']

### Apply Leiden Algorithm to Detect Communities

Choice of Leiden over Louvain:
- Can handle **directed graphs**, which is important in this case
- Typically yields more stable results across different runs compared to Louvain.
- Faster in practice than Louvain, especially on networks where the community structure is deeply nested.
- Can produce a hierarchy of partitions that are locally optimally partitioned.

**Choice of RBERVertexPartition**</br>
Model Basis: This partition type is based on the Erdős-Rényi Model but adjusted for degree sequences (hence, sometimes referred to in relation to the configuration model). The Erdős-Rényi model in its classic form assumes that all possible edges between nodes have an equal probability of existing, leading to a random network. The RBERVertexPartition adapts this concept to evaluate the significance of the observed community structure against a backdrop where connections are more randomly distributed but constrained by the degree sequence.

Use Case: Suitable for analyses where the emphasis is on understanding how the observed community divisions stand out against a null model of random connections with similar degree characteristics. It's particularly useful when the question revolves around the significance of the community structure rather than just its existence.

Reference: Chat-GPT, https://leidenalg.readthedocs.io/en/stable/reference.html

import leidenalg

# Find the optimal partition with the Leiden algorithm
partition = leidenalg.find_partition(g, leidenalg.RBERVertexPartition, weights='weight', n_iterations=-1, resolution_parameter=5)

# Get the community labels for each node
community_labels = partition.membership

# create a dataframe of the transaction_ids and community labels
coms_df = pd.DataFrame({'trxn_id': nodes, 'Community Label': community_labels})
coms_df

coms_df['Community Label'].nunique()

# Save Community Labels Dataframe to add to Customer Details

coms_df.to_csv('Community_labels.csv')
