==> SNA and GNNs/Code/5. Extract Community Features.ipynb <==
# 5. Extract Community Features

import pandas as pd
import numpy as np

# Import Data

# import new data after feature engineering
wires = pd.read_csv('wires.csv').iloc[:, 1:]
emts = pd.read_csv('emts.csv').iloc[:, 1:]
cash = pd.read_csv('cash.csv').iloc[:, 1:]
cust_info = pd.read_csv('cust_info.csv').iloc[:, 1:]
detail_cust = pd.read_csv('detailed_cust_info.csv').iloc[:, 1:]
ext_info = pd.read_csv('external_info.csv').iloc[:, 1:]
coms_df = pd.read_csv('Community_labels.csv').iloc[:, 1:]

# convert the amount in cash df to type float
cash['amount'] = cash['amount'].astype(float)

# Merge Community Labels to DataFrames

coms_df

# Append Community Labels to Transaction Data

# create a function to get the community labels for each customer's transactions
def get_labels(df):
    """
    """
    # Merge and keep needed columns for feature extraction by customer
    df1 = pd.merge(df, coms_df, how='left', on='trxn_id')[['trxn_id', 'sender_global_id', 'rec_global_id', 'Community Label']]
    
    # create the dataframe for each customer
    send_ls = df1[['sender_global_id', 'Community Label']].rename(columns={'sender_global_id': 'rec_global_id'})
    rec_ls = df1[['rec_global_id', 'Community Label']]
    
    # concat the two dataframes
    labels = pd.concat([send_ls, rec_ls]).rename(columns={'rec_global_id':'Global_id'})
    
    return labels

# Get the dataframe for wire transactions
wire_labels = get_labels(wires)
wire_labels

# Get the dataframe for emt transactions
emt_labels = get_labels(emts)
emt_labels

# Get labels for cash transactions (since only one customer no function needed)
cash_labels = pd.merge(cash, coms_df, how='left', on='trxn_id')[['Global_id', 'Community Label']]
cash_labels

# Feature Extraction using Group By

### Concat dataframes to single labels dataframe

# Concat the dataframes together
labels = pd.concat([wire_labels, emt_labels, cash_labels])
labels['Community Label'] = labels['Community Label'].astype(str)
labels

### Get Count Unique Communities by User

# Create a groupby to count the number of unique communities
# a user participated in
uniq_coms = pd.DataFrame(labels.groupby(['Global_id'])['Community Label'].nunique())
uniq_coms = uniq_coms.rename(columns={'Community Label': 'Num_unique_communities_participated'}).reset_index()
uniq_coms

### Get total number of communities participated

# Create a groupby to count the number of unique communities
# a user participated in
total_coms = pd.DataFrame(labels.groupby(['Global_id'])['Community Label'].count())
total_coms = total_coms.rename(columns={'Community Label': 'Total_communities_participated'}).reset_index()
total_coms

# Merge the dataframes

comm_fts = pd.merge(uniq_coms, total_coms, on='Global_id')
comm_fts

### Process Note

Initially we planned on adding a binary variable flag for each community to indicate whether or not an individual participated in the community. However, due to the large number of communities, this is not feasible we could not construct a matrix of that size due to memory constraints. 

# Save the DataFrame

comm_fts.to_csv('comm_fts.csv')
