==> SNA and GNNs/Code/6. Merge Extracted Features to Customers Data.ipynb <==
# 6. Merge Extract Features to Customers Data

import pandas as pd
import numpy as np
from sklearn.decomposition import PCA

# Import Data

# import new data after feature engineering
detailed_cust_info = pd.read_csv('detailed_cust_info.csv').iloc[:, 1:]
external_info = pd.read_csv('external_info.csv').iloc[:, 1:]

# Import extract community features data
comm_fts = pd.read_csv('comm_fts.csv').iloc[:, 1:]

# import participation data
wsw_part = pd.read_csv('wsw_part.csv').iloc[:, 1:]
wse_part = pd.read_csv('wse_part.csv').iloc[:, 1:]
wsc_part = pd.read_csv('wsc_part.csv').iloc[:, 1:]
ese_part = pd.read_csv('ese_part.csv').iloc[:, 1:]
esc_part = pd.read_csv('esc_part.csv').iloc[:, 1:]
esw_part = pd.read_csv('esw_part.csv').iloc[:, 1:]
csc_part = pd.read_csv('csc_part.csv').iloc[:, 1:]
csw_part = pd.read_csv('csw_part.csv').iloc[:, 1:]
cse_part = pd.read_csv('cse_part.csv').iloc[:, 1:]

# Merge Pariticipation Features

# create a list of the dataframes
parts = [wse_part, wsc_part, 
         ese_part, esc_part, esw_part, 
         csc_part, csw_part, cse_part]

# Merge the dataframs to one dataframe
parts_df = wsw_part.copy()

for i in range(len(parts)):
    parts_df = pd.merge(parts_df, parts[i], how='outer', on='Global_id', suffixes=(i, i+1))

# create a list of column names
part_cols = ['Global_id', 'wsw_part', 'wse_part', 'wsc_part', 
             'ese_part', 'esc_part', 'esw_part', 
             'csc_part', 'csw_part', 'cse_part']

# Set column names
parts_df.columns = part_cols

# Fill np.nans with 0 since nans appear after merging indicate
# an individual had participated in another network but didn't
# in the current one
parts_df = parts_df.fillna(0)
parts_df

### Merge pariticipation data to `Customer Details` DF

# Merge the participation data with the detailed_cust_info
# Given there are more nans, fill them with 0 (indicates no participation)
detail_cust = pd.merge(detailed_cust_info, parts_df, how='left', on='Global_id').fillna(0)
detail_cust

### Merge participation data to `External Info` DF

# Merge partitipcation df to external dataframe as well
ext_info = pd.merge(external_info, parts_df, how='left', on='Global_id').fillna(0)
ext_info

# Merge the Community Features

# Merge to detail_cust DF
detail_cust1 = pd.merge(detail_cust, comm_fts, how='left', on='Global_id').fillna(0)
detail_cust1

# Merge to ext_info DF
ext_info1 = pd.merge(ext_info, comm_fts, how='left', on='Global_id').fillna(0)
ext_info1

# Save Updated Dataframes

# Save
detail_cust1.to_csv('detailed_cust_info.csv')
ext_info1.to_csv('external_info.csv')
