==> SNA and GNNs/Code/7. Customers Graph.ipynb <==
# 7. Customers Graph

import pandas as pd
import numpy as np
import torch as th
import dgl

#!pip install dgl

# Import Data

# import new data after feature engineering
wires = pd.read_csv('wires.csv').iloc[:, 1:]
emts = pd.read_csv('emts.csv').iloc[:, 1:]
cash = pd.read_csv('cash.csv').iloc[:, 1:]
cust_info = pd.read_csv('cust_info.csv').iloc[:, 1:]
detail_cust = pd.read_csv('detailed_cust_info.csv').iloc[:, 1:]
ext_info = pd.read_csv('external_info.csv').iloc[:, 1:]

For the heterogeneous graph:

**Nodes**
- `customers` --> customers that are recorded in the kyc dataframe
- `external users` --> customers that are not recorded in the kyc dataframe
    
**Edges**
- `wire` transactions --> between customers and external customers
- `emt` transactions --> between customers and external customers
- `cash1` transactions --> customer self-loop
    
**Node Features**
- customers: detail_cust dataframe
- external customers: ext_info dataframe

**Edge Features**
- corresponding transaction data

#### Canonical Edge Types (COO form)

Note the edges are abbreviated as: **Source-transaction_type-Destination**
- e.g. ('customer', 'cwc', 'customer') --> customer-wire_trxn-customer --> 'cwc'

The types of connections we will make include (COO format):
- ('customer', 'cwc', 'customer')
- ('customer', 'cwe', 'external')
- ('external', 'ewc', 'customer')
- ('customer', 'cec', 'customer')
- ('customer', 'cee', 'external')
- ('external', 'eec', 'customer')
- ('customer', 'cc_self', 'customer')
- ('external', 'ee_self', 'external')

# Get Node Features for Customers

detail_cust

# Get Node Features for External Users

ext_info

# Get COO Tensors for Edges

#### Create maps to map the global ids to the corresponding index in the node features dataframe

# create a function to generate a map for a corresponding node features table
def gen_map(df, transactions):
    """
    """
    
    # If mapping transactions, get list of transaction ids
    if transactions:
        gids = list(df['trxn_id'])
    else:
        # get global_ids for indexes of customers
        gids = list(df['Global_id'])
    
    # create a list for the indexes to map to
    indexes = [ x for x in range(len(df)) ]

    # create a dictionary for the mapped indexes
    i_map = {gid:index for gid, index in zip(gids, indexes)}
    
    return i_map

# Create map for customers
cust_map = gen_map(cust_info, False)

# Create map for external users
ext_map = gen_map(ext_info, False)

#### Create functions to extract COO format

# create a function to get edges with the corresponding index values in the node features dataframe
def get_edges(df, ext_source, ext_dest):
    """
    """
    edges_df = pd.DataFrame()
    
    edges_df['trxn_id'] = df['trxn_id']
    
    # Map the global ids to the corresponding index in the customer node features dataframe
    edges_df['Source'] = df['sender_global_id'].map(cust_map)
    edges_df['Destination'] = df['rec_global_id'].map(cust_map)
    
    # If source/dest is external map to indexes in external users node features dataframe
    if ext_source:
        edges_df['Source'] = df['sender_global_id'].map(ext_map)
    elif ext_dest:
        edges_df['Destination'] = df['rec_global_id'].map(ext_map)
    
    return edges_df

# create a function to store the edges dataframe as a tuple with tensors
def get_tensors(edges, self_loop):
    """
    """
    
    if not self_loop:
        filtered = edges.drop(columns='trxn_id')
    else:
        filtered = edges
        
    # transpose the dataframe to get the format for the tensors
    values = filtered.values.transpose()
    
    # convert the values to tensors
    u = th.tensor(values[0].copy())
    v = th.tensor(values[1].copy())
    
    # return a tuple of node_tensors
    return (u, v)

## Get edges for wire transactions

# Get info for wire transactions
w_trxns = wires[['trxn_id', 'sender_global_id', 'rec_global_id', 'sender_ext', 'receiver_ext']]
w_trxns

### Extract COO for ('customer', 'wire trxn', 'customer') 

# find transactions where both customers are not external users
cwc = w_trxns[(w_trxns['sender_ext'] == 0) & (w_trxns['receiver_ext'] == 0)][['trxn_id', 'sender_global_id', 'rec_global_id']]

# Extract the edges
cwc_edges = get_edges(cwc, False, False)
cwc_edges

# get the node-tensors for these edges
cwc_ten = get_tensors(cwc_edges, False)
cwc_ten

### Extract COO for ('customer', 'wire trxn', 'external') 

# find transactions where receiver is an external user
cwe = w_trxns[(w_trxns['sender_ext'] == 0) & (w_trxns['receiver_ext'] == 1)][['trxn_id', 'sender_global_id', 'rec_global_id']]

# Extract the edges
cwe_edges = get_edges(cwe, False, True)
cwe_edges

# get the node-tensors for these edges
cwe_ten = get_tensors(cwe_edges, False)
cwe_ten

### Extract COO for ('external', 'wire trxn', 'customer') 

# find transactions where sender is an external user
ewc = w_trxns[(w_trxns['sender_ext'] == 1) & (w_trxns['receiver_ext'] == 0)][['trxn_id', 'sender_global_id', 'rec_global_id']]

# Extract the edges
ewc_edges = get_edges(ewc, True, False)
ewc_edges

# get the node-tensors for these edges
ewc_ten = get_tensors(ewc_edges, False)
ewc_ten

## Get edges for emt transactions

# Get info for wire transactions
e_trxns = emts[['trxn_id', 'sender_global_id', 'rec_global_id', 'sender_ext', 'receiver_ext']]
e_trxns

### Extract COO for ('customer', 'emt trxn', 'customer') 

# find transactions where both customers are not external users
cec = e_trxns[(e_trxns['sender_ext'] == 0) & (e_trxns['receiver_ext'] == 0)][['trxn_id', 'sender_global_id', 'rec_global_id']]

# Extract the edges
cec_edges = get_edges(cec, False, False)
cec_edges

# get the node-tensors for these edges
cec_ten = get_tensors(cec_edges, False)
cec_ten

### Extract COO for ('customer', 'emt trxn', 'external') 

# find transactions where receiver is an external user
cee = e_trxns[(e_trxns['sender_ext'] == 0) & (e_trxns['receiver_ext'] == 1)][['trxn_id', 'sender_global_id', 'rec_global_id']]

# Extract the edges
cee_edges = get_edges(cee, False, True)
cee_edges

# get the node-tensors for these edges
cee_ten = get_tensors(cee_edges, False)
cee_ten

### Extract COO for ('external', 'emt trxn', 'customer') 

# find transactions where sender is an external user
eec = e_trxns[(e_trxns['sender_ext'] == 1) & (e_trxns['receiver_ext'] == 0)][['trxn_id', 'sender_global_id', 'rec_global_id']]

# Extract the edges
eec_edges = get_edges(eec, True, False)
eec_edges

# get the node-tensors for these edges
eec_ten = get_tensors(eec_edges, False)
eec_ten

# Create self loops for cash transactions

### Create self loops for customer nodes

# Create self loops for customer nodes
gids = [ x for x in range(len(cust_info['Global_id']))]
cust_self = pd.DataFrame({'Global_id': gids, 'Gids': gids})
cust_self

# get the tensors
cself_ten = get_tensors(cust_self, True)
cself_ten

### Create self loops for external nodes

# Create self loops for customer nodes
gids = [ x for x in range(len(ext_info['Global_id']))]
ext_self = pd.DataFrame({'Global_id': gids, 'Gids': gids})
ext_self

# get the tensors
extself_ten = get_tensors(ext_self, True)
extself_ten

# Construct Graph

#### Create graph data using edge COOs

# Combine the edge COOs into the graph data
graph_data = {
    ('customer', 'cwc', 'customer'): cwc_ten,
    ('customer', 'cwe', 'external'): cwe_ten,
    ('external', 'ewc', 'customer'): ewc_ten,
    ('customer', 'cec', 'customer'): cec_ten,
    ('customer', 'cee', 'external'): cee_ten,
    ('external', 'eec', 'customer'): eec_ten,
    ('customer', 'cc_self', 'customer'): cself_ten,
    ('external', 'ee_self', 'external'): extself_ten,   
}

g = dgl.heterograph(graph_data)

g.ntypes

g.etypes

g.canonical_etypes

g.num_nodes()

g.num_edges()

# Add node/edge features

# create a function to extract feature data into tensors
def get_fts(features_df):
    """
    """
    # convert the dataframe to numpy 
    x = features_df.to_numpy()
    
    # Create a tensor with feature data 
    tens = th.tensor(x.copy()).float()
    
    return tens

### Add customer node features

# Drop the id column in cust_info
cust_fts = detail_cust.drop(columns='Global_id')
cust_fts

# create a tensor with customer feature data
x_custs = get_fts(cust_fts)
x_custs

# Add features for customer nodes
g.nodes['customer'].data['features'] = x_custs

### Add external user node features

# Drop the id column in ext_info
ext_fts = ext_info.drop(columns='Global_id')
ext_fts

# create a tensor with external user feature data
x_ext = get_fts(ext_fts)
x_ext

# Add features for customer nodes
g.nodes['external'].data['features'] = x_ext

# Get edge features

Extract the edge features for each canonical edge type.

# Create a function to extract the features for a given canonical edge type
def get_efeats(edges, trxns_df):
    """
    """
    
    # Generate a map to map the corresponding transactions to the indexes in the trxns_df
    i_map = gen_map(edges, True)
    
    # Get a list of transaction ids in the edges df
    ids = list(edges['trxn_id'])
    
    # filter for the trxns that have ids in the edges df
    selected = trxns_df[trxns_df['trxn_id'].isin(ids)]
    
    # Add a column for the corresponding index in trxns_df
    selected.loc[:, 'index'] = selected.loc[:, 'trxn_id'].map(i_map)
    
    # order the dataframe by the index and drop unecessary columns
    selected = selected.sort_values(by='index')
    selected = selected.drop(columns=['trxn_id', 'index'])
    
    # Get the tensors for the data
    x = get_fts(selected)
    
    return x

### Get ('customer', 'wire trxn', 'customer') features

# Get wire transactions related to cwc
cwc_fts = get_efeats(cwc_edges, wires)
cwc_fts

### Get ('customer', 'wire trxn', 'external') features

# Get wire transactions related to cwc
cwe_fts = get_efeats(cwe_edges, wires)
cwe_fts

### Get ('external', 'wire trxn', 'customer') features

# Get wire transactions related to cwc
ewc_fts = get_efeats(ewc_edges, wires)
ewc_fts

### Get ('customer', 'emt trxn', 'customer') features

# Get wire transactions related to cwc
cec_fts = get_efeats(cec_edges, emts)
cec_fts

### Get ('customer', 'emt trxn', 'external') features

# Get wire transactions related to cwc
cee_fts = get_efeats(cee_edges, emts)
cee_fts

### Get ('external', 'emt trxn', 'customer') features

# Get wire transactions related to cwc
eec_fts = get_efeats(eec_edges, emts)
eec_fts

### Get ('customer', 'self_loop', 'customer') Cash Transaction features

# since the cash transactions are stored in the cust_info
# we do not need to remap the indexes as they are in the right order
c_trxns = cust_info.iloc[:, -4:]
c_trxns

# Get the tensors
cself_fts = get_fts(c_trxns)
cself_fts

### Get ('external', 'self_loop', 'external') Cash Transaction features

# since the external users have 0 transactions, create tensors with just 0s
eself_fts = th.zeros((len(ext_info)), 4)
eself_fts 

## Set edge features

# set the edge features in the graph
g.edges[('customer', 'cwc', 'customer')].data['features'] = cwc_fts
g.edges[('customer', 'cwe', 'external')].data['features'] = cwe_fts
g.edges[('external', 'ewc', 'customer')].data['features'] = ewc_fts
g.edges[('customer', 'cec', 'customer')].data['features'] = cec_fts
g.edges[('customer', 'cee', 'external')].data['features'] = cee_fts
g.edges[('external', 'eec', 'customer')].data['features'] = eec_fts
g.edges[('customer', 'cc_self', 'customer')].data['features'] = cself_fts
g.edges[('external', 'ee_self', 'external')].data['features'] = eself_fts

g

# Save graph

# Save the graph to a file
dgl.save_graphs('customers_graph', [g])
